Replace high, low by single digits array
High, low doesn't scale well in terms of memory to more than two
radix steps. Here we now read the keys again in each radix step
and overwrite a digits array at the beginning of the step. This
reduces memory. It even reduces instructions because allocation
of the additional array is probably more costly than accessing
the keys twice.